server/requirements.txt
Flask
flask-pymongo
flask-bcrypt
flask-login
python-dotenv
